+++ src/dgisim/card/cards_set.py
@@ -13,6 +13,7 @@ _DEFAULT_CARDS: list[type[Card]] = [
     ColdBloodedStrike,
     DescentOfDivinity,
     EmbraceOfWinds,
+    GalesOfReverie,
     GrandExpectation,
     IGotYourBack,
     KeenSight,